Assignment - 3
ack and his three friends have decided to go for a trip by sharing the expenses of the fuel equally. 

Write a Python program to calculate the amount (in Rs) each of them need to put in for the complete (both to and fro) journey.
The program should also display True, if the amount to be paid by each person is divisible by 5, otherwise it should display False. (Hint: Use the relational operators in print statement.)

Assume that mileage of the vehicle, amount per litre of fuel and distance for one way are given. 

Test your code by using the given sample inputs.