Assignment - 5
Write a pseudo-code to find out if a given year is a leap year or not.
Any year which is divisible by 4 and not by 100 are leap years. Otherwise, any year which is divisible by 400 is also a leap year.
Use the drag and drop option provided to create the correct pseudo-code.