Exercise - 8
Write a Go program to calculate and display the final fees to be paid by a student.

The students are entitled to get scholarship based on the marks scored in the qualifying exam. Scholarship percentage should be considered as half of the marks scored by the student on the course fee. Apart from the course fee (after deducting the scholarship amount), the students have to pay an extra service fees.

Consider that course fee, marks scored by the student and extra fees amount are given. Assume that the marks is out of 100 and need not always be integer.

Test your code by using the given sample inputs.